Practical 2

+++ PR1/Problem_01.c
@@ -7,7 +7,7 @@ int main() {
     printf("Sum of A and B is: %d\n", a+b);
     printf("Sub of A and B is: %d\n", a-b);
     printf("Product of A and B is: %d\n", a*b);
-    printf("Division of A and B is: %d\n", a/b);
+    printf("Division of A and B is: %.2f\n", (float)a/b);
 
     return 0;
 }
